requirements.txt
os
Fernet
python-cryptography
